# Restructuring to allow preprocessing, tracking and rendering to be compiled separately 