--- a/se_denseslam/include/se/DenseSLAMSystem.h
+++ b/se_denseslam/include/se/DenseSLAMSystem.h
@@ -52,10 +52,6 @@
 #include <se/preprocessing.hpp>
 #include <se/tracking.hpp>
 
-/*
- * Use SE_FIELD_TYPE macro to define the DenseSLAMSystem instance.
- */
-typedef SE_FIELD_TYPE FieldType;
 template <typename T>
 using Volume = VolumeTemplate<T, se::Octree>;