--- a/se_denseslam/src/DenseSLAMSystem.cpp
+++ b/se_denseslam/src/DenseSLAMSystem.cpp
@@ -41,7 +41,6 @@
 #include <se/vtk-io.h>
 #include "timings.h"
 #include <perfstats.h>
-#include "tracking.cpp"
 #include "rendering.cpp"
 #include "bfusion/mapping_impl.hpp"
 #include "kfusion/mapping_impl.hpp"