Changed the project structure so that preprocessing, tracking and rendering are split into hpp and cpp files. This should make code separation cleaner and allow faster compilation for incremental changes.

### Preprocessing
Split into an hpp and cpp and added explicit instantiation of template function `vertex2normalKernel()`. The template was for a `bool` so all possible functions have been instantiated.

### Tracking
Just split into an hpp and cpp. No notable changes.

### Structural changes
- Moved the `SE_FIELD_TYPE` `typedef` from `DenseSLAMSystem.h` to `volume_traits.hpp ` so that it will be available to all headers (rendering.hpp etc).
- Moved the `bfusion` and `kfusion` folders from `src` to `include`.
- Some bug fixes and code style improvements.

### Rendering
Split into hpp and cpp. Two functions were split fine (`renderDepthKernel()` and `renderTrackKernel()`) but the other two (`raycastKernel()` and `renderVolumeKernel()`) were templated andI left them in the header for now. I am not particularly happy about this but maybe it would be better to use the new file structure as early as possible and make rendering cleaner later.

### Benchmarks
I run the generated benchmarks for both the SDF and occupancy mapping representations for the old and new code. I don't think the changes had an impact in performance, although the benchmarking wasn't that thorough. Each benchmark was run 10 times and the results are below. What is strange is that in both the new and old code the time to complete the SDF benchmark varied a lot. No clue what might be causing this at the moment.

New code
```bash
# SDF TUM
# Benchmarking './build/se_apps/se-denseslam-sdf-benchmark -i /home/srl/Documents/Datasets/TUM_RGBD/rgbd_dataset_freiburg1_desk_mrcnn/scene.raw -s 6 -p 0.5,0.5,0.5 -z 4 -c 2 -r 1 -k 525,525,319.5,239.5'
# for 10 iterations
-----------------------------------
Mean:                 22.928244 s
Standard deviation:   12.319538 s
Min:                  8.892734 s
Max:                  38.106689 s
Total time:           229.282444 s

# OFUSION TUM
# Benchmarking './build/se_apps/se-denseslam-ofusion-benchmark -i /home/srl/Documents/Datasets/TUM_RGBD/rgbd_dataset_freiburg1_desk_mrcnn/scene.raw -s 6 -p 0.5,0.5,0.5 -z 4 -c 2 -r 1 -k 525,525,319.5,239.5'
# for 10 iterations
-----------------------------------
Mean:                 23.038625 s
Standard deviation:   0.911637 s
Min:                  20.769309 s
Max:                  24.296465 s
Total time:           230.386250 s
```

Old code
```bash
# SDF TUM
# Benchmarking './build/se_apps/se-denseslam-sdf-benchmark -i /home/srl/Documents/Datasets/TUM_RGBD/rgbd_dataset_freiburg1_desk_mrcnn/scene.raw -s 6 -p 0.5,0.5,0.5 -z 4 -c 2 -r 1 -k 525,525,319.5,239.5'
# for 10 iterations
-----------------------------------
Mean:                 22.968320 s
Standard deviation:   12.656391 s
Min:                  8.807932 s
Max:                  39.406392 s
Total time:           229.683196 s

# OFUSION TUM
# Benchmarking './build/se_apps/se-denseslam-ofusion-benchmark -i /home/srl/Documents/Datasets/TUM_RGBD/rgbd_dataset_freiburg1_desk_mrcnn/scene.raw -s 6 -p 0.5,0.5,0.5 -z 4 -c 2 -r 1 -k 525,525,319.5,239.5'
# for 10 iterations
-----------------------------------
Mean:                 25.890067 s
Standard deviation:   4.092455 s
Min:                  21.455162 s
Max:                  34.664623 s
Total time:           258.900666 s
```