@sotpapathe 2019-03-25
More benchmarks

#### New code (7e48804a1676841920c29523976adc3428675395)

| dataset | noise\_factor | ATE | preprocessing | integration | raycasting | computation |
| :- | :-: | :-: | :-: | :-: | :-: | :-: |
| `liv_traj_0` | 0.1000 | 0.0074 | 0.0001 | 0.0123 | 0.0133 | 0.0290 |
| `liv_traj_0` | 0.1000 | 0.0073 | 0.0001 | 0.0122 | 0.0133 | 0.0289 |
| `liv_traj_0` | 0.1000 | 0.0074 | 0.0001 | 0.0153 | 0.0166 | 0.0359 |
| `liv_traj_0` | 0.1000 | 0.0072 | 0.0001 | 0.0124 | 0.0135 | 0.0292 |

| dataset | noise\_factor | ATE | preprocessing | integration | raycasting | computation |
| :- | :-: | :-: | :-: | :-: | :-: | :-: |
| `liv_traj_1` | 0.1000 | 0.0112 | 0.0012 | 0.0374 | 0.0404 | 0.1019 |
| `liv_traj_1` | 0.1000 | 0.0113 | 0.0001 | 0.0093 | 0.0108 | 0.0230 |
| `liv_traj_1` | 0.1000 | 0.0112 | 0.0004 | 0.0374 | 0.0444 | 0.0938 |
| `liv_traj_1` | 0.1000 | 0.0112 | 0.0008 | 0.0137 | 0.0131 | 0.0336 |

| dataset | noise\_factor | ATE | preprocessing | integration | raycasting | computation |
| :- | :-: | :-: | :-: | :-: | :-: | :-: |
| `liv_traj_2` | 0.1000 | 0.0037 | 0.0003 | 0.0207 | 0.0188 | 0.0446 |
| `liv_traj_2` | 0.1000 | 0.0037 | 0.0002 | 0.0152 | 0.0137 | 0.0329 |
| `liv_traj_2` | 0.1000 | 0.0038 | 0.0002 | 0.0271 | 0.0240 | 0.0565 |
| `liv_traj_2` | 0.1000 | 0.0038 | 0.0005 | 0.0173 | 0.0141 | 0.0365 |

| dataset | noise\_factor | ATE | preprocessing | integration | raycasting | computation |
| :- | :-: | :-: | :-: | :-: | :-: | :-: |
| `liv_traj_3` | 0.1000 | 0.7451 | 0.0001 | 0.0141 | 0.0156 | 0.0345 |
| `liv_traj_3` | 0.1000 | 0.6402 | 0.0002 | 0.0069 | 0.0118 | 0.0226 |
| `liv_traj_3` | 0.1000 | 0.6139 | 0.0001 | 0.0130 | 0.0157 | 0.0331 |
| `liv_traj_3` | 0.1000 | 1.2634 | 0.0003 | 0.0049 | 0.0092 | 0.0170 |



#### Old code (devel 0a90a2a2dc37a6ec18113dd4e56f0b45770830be)

| dataset | noise\_factor | ATE | preprocessing | integration | raycasting | computation |
| :- | :-: | :-: | :-: | :-: | :-: | :-: |
| `liv_traj_0` | 0.1000 | 0.0079 | 0.0003 | 0.0183 | 0.0188 | 0.0440 |
| `liv_traj_0` | 0.1000 | 0.0079 | 0.0001 | 0.0163 | 0.0176 | 0.0381 |
| `liv_traj_0` | 0.1000 | 0.0079 | 0.0001 | 0.0153 | 0.0165 | 0.0357 |
| `liv_traj_0` | 0.1000 | 0.0076 | 0.0001 | 0.0145 | 0.0153 | 0.0336 |

| dataset | noise\_factor | ATE | preprocessing | integration | raycasting | computation |
| :- | :-: | :-: | :-: | :-: | :-: | :-: |
| `liv_traj_1` | 0.1000 | 0.0112 | 0.0007 | 0.0442 | 0.0388 | 0.0967 |
| `liv_traj_1` | 0.1000 | 0.0113 | 0.0004 | 0.0419 | 0.0373 | 0.0895 |
| `liv_traj_1` | 0.1000 | 0.0112 | 0.0005 | 0.0376 | 0.0344 | 0.0834 |
| `liv_traj_1` | 0.1000 | 0.0111 | 0.0001 | 0.0133 | 0.0118 | 0.0282 |

| dataset | noise\_factor | ATE | preprocessing | integration | raycasting | computation |
| :- | :-: | :-: | :-: | :-: | :-: | :-: |
| `liv_traj_2` | 0.1000 | 0.0036 | 0.0002 | 0.0271 | 0.0209 | 0.0524 |
| `liv_traj_2` | 0.1000 | 0.0036 | 0.0002 | 0.0230 | 0.0178 | 0.0447 |
| `liv_traj_2` | 0.1000 | 0.0036 | 0.0002 | 0.0251 | 0.0195 | 0.0494 |
| `liv_traj_2` | 0.1000 | 0.0036 | 0.0001 | 0.0178 | 0.0138 | 0.0346 |

| dataset | noise\_factor | ATE | preprocessing | integration | raycasting | computation |
| :- | :-: | :-: | :-: | :-: | :-: | :-: |
| `liv_traj_3` | 0.1000 | 1.1698 | 0.0001 | 0.0075 | 0.0095 | 0.0206 |
| `liv_traj_3` | 0.1000 | 1.2615 | 0.0002 | 0.0048 | 0.0104 | 0.0182 |
| `liv_traj_3` | 0.1000 | 1.2629 | 0.0001 | 0.0057 | 0.0126 | 0.0207 |
| `liv_traj_3` | 0.1000 | 0.7542 | 0.0002 | 0.0129 | 0.0123 | 0.0293 |